Crate pallet_moderation

Moderation Module

The Moderation module allows any user (account) to report an account, space, post or even IPFS CID, if they think it’s a spam, abuse or inappropriate for a specific space.

Moderators of a space can review reported entities and suggest a moderation status for them: Block or Allowed. A space owner can make a final decision: either block or allow any entity within the space they control.

This pallet also has a setting to auto-block the content after a specific number of statuses from moderators that suggest to block the entity. If the entity is added to allow list, then the entity cannot be blocked.

The next rules applied to the blocked entities:

A post cannot be added to a space if an IPFS CID of this post is blocked in this space.

An account cannot create posts in a space if this account is blocked in this space.

Modules

functions

Structs

ModerationSettings
A custom moderation settings for a certain space (key).

Module
The module declaration.

NextReportId
The next moderation report id.

Report

ReportById
Report details by its id (key).

ReportIdByAccount
Report id if entity (key 1) was reported by a specific account (key 2)

ReportIdsByEntityInSpace
Ids of all reports related to a specific entity (key 1) sent to this space (key 2).

ReportIdsBySpaceId
Ids of all reports in this space (key).

SpaceModerationSettings

SpaceModerationSettingsUpdate

StatusByEntityInSpace
An entity (key 1) status (Blocked or Allowed) in this space (key 2).

SuggestedStatus

SuggestedStatusesByEntityInSpace
Entity (key 1) statuses suggested by space (key 2) moderators.

Enums

Call
Dispatchable calls.

EntityId

EntityStatus
Entity status is used in two cases: when moderators suggest a moderation status for a reported entity; or when a space owner makes a final decision to either block or allow this entity within the space.

Error

RawEvent
Events for this module.

Constants

FIRST_REPORT_ID

Traits

Trait
The pallet’s configuration trait.

Type Definitions

Event
RawEvent specialized for the configuration Trait

ReportId